[management] Move validator-config to its own file
Refactor commit -- nothing interesting here...
