+++ config/management/src/lib.rs
@@ -5,23 +5,13 @@
 
 mod error;
 mod secure_backend;
+mod validator_config;
 
-use crate::{error::Error, secure_backend::SecureBackend};
-use libra_crypto::{ed25519::Ed25519PublicKey, hash::CryptoHash, x25519, ValidCryptoMaterial};
-use libra_network_address::{NetworkAddress, RawNetworkAddress};
+use crate::{error::Error, secure_backend::SecureBackend, validator_config::ValidatorConfig};
+use libra_crypto::ed25519::Ed25519PublicKey;
 use libra_secure_storage::{Storage, Value};
-use libra_secure_time::{RealTimeService, TimeService};
-use libra_types::{
-    account_address::AccountAddress,
-    transaction::{RawTransaction, SignedTransaction, Transaction},
-    waypoint::Waypoint,
-};
-use std::{
-    convert::{TryFrom, TryInto},
-    fmt::Write,
-    str::FromStr,
-    time::Duration,
-};
+use libra_types::{transaction::Transaction, waypoint::Waypoint};
+use std::{convert::TryInto, fmt::Write, str::FromStr};
 use structopt::StructOpt;
 
 pub mod constants {
@@ -116,75 +106,7 @@ impl Command {
 
     pub fn validator_config(self) -> Result<Transaction, Error> {
         if let Command::ValidatorConfig(config) = self {
-            let mut storage: Box<dyn Storage> = config.backend.backend.try_into()?;
-            if !storage.available() {
-                return Err(Error::LocalStorageUnavailable);
-            }
-
-            let consensus_key = storage
-                .get_public_key(constants::CONSENSUS_KEY)
-                .map_err(|e| Error::LocalStorageReadError(e.to_string()))?
-                .public_key;
-            let fullnode_network_key = storage
-                .get_public_key(constants::FULLNODE_NETWORK_KEY)
-                .map_err(|e| Error::LocalStorageReadError(e.to_string()))?
-                .public_key;
-            let validator_network_key = storage
-                .get_public_key(constants::VALIDATOR_NETWORK_KEY)
-                .map_err(|e| Error::LocalStorageReadError(e.to_string()))?
-                .public_key;
-
-            let fullnode_address = RawNetworkAddress::try_from(&config.fullnode_address)
-                .expect("Unable to serialize fullnode network address from config");
-            let validator_address = RawNetworkAddress::try_from(&config.validator_address)
-                .expect("Unable to serialize validator network address from config");
-
-            let fullnode_network_key = fullnode_network_key.to_bytes();
-            let fullnode_network_key: x25519::PublicKey = fullnode_network_key
-                .as_ref()
-                .try_into()
-                .expect("Unable to decode x25519 from fullnode_network_key");
-
-            let validator_network_key = validator_network_key.to_bytes();
-            let validator_network_key: x25519::PublicKey = validator_network_key
-                .as_ref()
-                .try_into()
-                .expect("Unable to decode x25519 from validator_network_key");
-
-            let owner_key = storage
-                .get_public_key(constants::OWNER_KEY)
-                .map_err(|e| Error::LocalStorageReadError(e.to_string()))?
-                .public_key;
-
-            // TODO(davidiw): The signing key, parameter 2, will be deleted soon, so this is a
-            // temporary hack to reduce over-engineering.
-            let script = transaction_builder::encode_register_validator_script(
-                consensus_key.to_bytes().to_vec(),
-                owner_key.to_bytes().to_vec(),
-                validator_network_key.to_bytes(),
-                validator_address.into(),
-                fullnode_network_key.to_bytes(),
-                fullnode_address.into(),
-            );
-
-            let sender = config.owner_address;
-            // TODO(davidiw): In genesis this is irrelevant -- afterward we need to obtain the
-            // current sequence number by querying the blockchain.
-            let sequence_number = 0;
-            let expiration_time = RealTimeService::new().now() + constants::TXN_EXPIRATION_SECS;
-            let raw_transaction = RawTransaction::new_script(
-                sender,
-                sequence_number,
-                script,
-                constants::MAX_GAS_AMOUNT,
-                constants::GAS_UNIT_PRICE,
-                Duration::from_secs(expiration_time),
-            );
-            let signature = storage
-                .sign_message(constants::OWNER_KEY, &raw_transaction.hash())
-                .map_err(|e| Error::LocalStorageSigningError(e.to_string()))?;
-            let signed_txn = SignedTransaction::new(raw_transaction, owner_key, signature);
-            Ok(Transaction::UserTransaction(signed_txn))
+            config.execute()
         } else {
             let expected = CommandName::ValidatorConfig.to_string();
             let actual = CommandName::from(&self).to_string();
@@ -331,19 +253,6 @@ pub struct SingleBackend {
     backend: SecureBackend,
 }
 
-// TODO(davidiw) add operator_address, since that will eventually be the identity producing this.
-#[derive(Debug, StructOpt)]
-pub struct ValidatorConfig {
-    #[structopt(long)]
-    owner_address: AccountAddress,
-    #[structopt(long)]
-    validator_address: NetworkAddress,
-    #[structopt(long)]
-    fullnode_address: NetworkAddress,
-    #[structopt(flatten)]
-    backend: SingleBackend,
-}
-
 /// These tests depends on running Vault, which can be done by using the provided docker run script
 /// in `docker/vault/run.sh`.
 /// Note: Some of these tests may fail if you run them too quickly one after another due to data
@@ -352,8 +261,9 @@ pub struct ValidatorConfig {
 #[cfg(test)]
 pub mod tests {
     use super::*;
+    use libra_network_address::NetworkAddress;
     use libra_secure_storage::{Policy, Value, VaultStorage};
-    use libra_types::transaction::TransactionPayload;
+    use libra_types::{account_address::AccountAddress, transaction::TransactionPayload};
 
     const VAULT_HOST: &str = "http://localhost:8200";
     const VAULT_ROOT_TOKEN: &str = "root_token";
